refactor: modularize the outputs package
