+++ src/outputs/index.ts
@@ -1,12 +1,9 @@
 import type { error, Outputter } from "../types";
+import type { OptimizedProjectStats } from "./types";
 
-import {
-  EVENT_PULL_REQUEST,
-  EVENT_PUSH,
-  EVENT_REPOSITORY_DISPATCH,
-  EVENT_SCHEDULE,
-  EVENT_WORKFLOW_DISPATCH,
-} from "./constants";
+import { getOutputNamesFor } from "./names";
+import { getValuesForOutputs } from "./values";
+import { writeOutputs } from "./write";
 
 interface Context {
   readonly eventName: string;
@@ -16,63 +13,18 @@ interface Environment {
   readonly context: Context;
 }
 
-interface OptimizedProjectStats {
-  readonly optimizedCount: number;
-  readonly svgCount: number;
-}
-
 interface Params {
   readonly data: OptimizedProjectStats;
   readonly env: Environment;
   readonly out: Outputter;
 }
 
-type DataToOutput = (data: OptimizedProjectStats) => string;
-
-const enum OutputName {
-  DID_OPTIMIZE = "DID_OPTIMIZE",
-  OPTIMIZED_COUNT = "OPTIMIZED_COUNT",
-  SVG_COUNT = "SVG_COUNT",
-}
-
-const outputsMap: Map<OutputName, DataToOutput> = new Map([
-  [
-    OutputName.DID_OPTIMIZE,
-    (data) => `${data.optimizedCount > 0}`,
-  ],
-  [
-    OutputName.OPTIMIZED_COUNT,
-    (data) => `${data.optimizedCount}`,
-  ],
-  [
-    OutputName.SVG_COUNT,
-    (data) => `${data.svgCount}`,
-  ],
-]);
-
-function getOutputNamesFor(event: string): [OutputName[], error] {
-  switch (event) {
-  case EVENT_PULL_REQUEST:
-  case EVENT_PUSH:
-  case EVENT_REPOSITORY_DISPATCH:
-  case EVENT_SCHEDULE:
-  case EVENT_WORKFLOW_DISPATCH:
-  default:
-    return [[
-      OutputName.DID_OPTIMIZE,
-      OutputName.OPTIMIZED_COUNT,
-      OutputName.SVG_COUNT,
-    ], null];
-  }
-}
-
 function Set({ env, data, out }: Params): error {
-  const [names, err] = getOutputNamesFor(env.context.eventName);
-  for (const name of names) {
-    const fn = outputsMap.get(name) as DataToOutput;
-    const value = fn(data);
-    out.setOutput(name, value);
-  }
+  const { eventName } = env.context;
+
+  const [names, err] = getOutputNamesFor(eventName);
+  const outputNameToValue = getValuesForOutputs(names, data);
+  writeOutputs(out, outputNameToValue);
 
   return err;
 }
