+++ src/outputs/types.d.ts
@@ -1,0 +1,8 @@
+interface OptimizedProjectStats {
+  readonly optimizedCount: number;
+  readonly svgCount: number;
+}
+
+export type {
+  OptimizedProjectStats,
+};
